refactor(core, yaml): hoist the line-break rule out of yaml onto text
#341 collapsed roughly twenty open-coded LF/CR/CRLF chains in src/yaml onto
one `yaml::line_break` module, on the grounds that duplicated predicates
diverge silently (#106). #228 then wrote a fresh copy of the same rule into
`text::LineIndex`, justified in the module doc by "it indexes JSON and DSV
text too, where the YAML spelling does not apply".

That justification does not hold: the rule is byte-identical for all three
formats. The real obstacle was visibility — `line_break_len` is `pub(super)`
inside `src/yaml`. So the copy was #341's duplication reintroduced one module
over, with nothing asserting the two agreed.

Move the three helpers and their tests to `text::line_break` as `pub(crate)`.
`yaml::line_break` becomes a re-export and keeps the YAML-facing rationale —
the §5.4 citation, the `skip_line_break` exception, the SIMD carve-out — so
YAML call sites are untouched. `LineIndex::build`'s scan loop drops its
hand-rolled `match` for `line_break_len`, which also makes the CRLF case one
branch instead of a nested conditional.

Behaviour is unchanged: the corpus-shape golden, which records the retained
line-index bytes for every corpus file, is byte-identical.

Refs #228, #341
